@@ -18,21 +18,18 @@
  * limitations under the License.
  */
 
-import parts = require("./parts/index");
+import * as parts from "./parts/index";
 
-import globalSettings = require("./settings");
+import { debugMode } from "./settings";
 
-import ASS = require("./types/ass");
-import Style = require("./types/style");
-import Dialogue = require("./types/dialogue");
-import types = require("./types/misc");
-import Property = types.Property;
-import TypedTemplate = types.TypedTemplate;
+import { ASS } from "./types/ass";
+import { Style } from "./types/style";
+import { Dialogue } from "./types/dialogue";
+import { Property, TypedTemplate } from "./types/misc";
 
-import map = require("./utility/map");
+import { Map } from "./utility/map";
 
-import promise = require("./utility/promise");
-import DeferredPromise = promise.DeferredPromise;
+import { Promise, DeferredPromise } from "./utility/promise";
 
 /**
  * An interface for a stream.
@@ -41,7 +38,7 @@ export interface Stream {
 	/**
 	 * @return {!Promise.<?string>} A promise that will be resolved with the next line, or null if the stream is exhausted.
 	 */
-	nextLine(): promise.Promise<string>;
+	nextLine(): Promise<string>;
 }
 
 /**
@@ -57,22 +54,22 @@ export class StringStream implements Stream {
 	/**
 	 * @return {!Promise.<?string>} A promise that will be resolved with the next line, or null if the string has been completely read.
 	 */
-	nextLine(): promise.Promise<string> {
-		var result: promise.Promise<string> = null;
+	nextLine(): Promise<string> {
+		var result: Promise<string> = null;
 
 		if (this._readTill < this._str.length) {
 			var nextNewLinePos = this._str.indexOf("\n", this._readTill);
 			if (nextNewLinePos !== -1) {
-				result = promise.Promise.resolve(this._str.substring(this._readTill, nextNewLinePos));
+				result = Promise.resolve(this._str.substring(this._readTill, nextNewLinePos));
 				this._readTill = nextNewLinePos + 1;
 			}
 			else {
-				result = promise.Promise.resolve(this._str.substr(this._readTill));
+				result = Promise.resolve(this._str.substr(this._readTill));
 				this._readTill = this._str.length;
 			}
 		}
 		else {
-			result = promise.Promise.resolve<string>(null);
+			result = Promise.resolve<string>(null);
 		}
 
 		return result;
@@ -96,7 +93,7 @@ export class XhrStream implements Stream {
 	/**
 	 * @return {!Promise.<?string>} A promise that will be resolved with the next line, or null if the stream is exhausted.
 	 */
-	nextLine(): promise.Promise<string> {
+	nextLine(): Promise<string> {
 		if (this._pendingDeferred !== null) {
 			throw new Error("XhrStream only supports one pending unfulfilled read at a time.");
 		}
@@ -177,14 +174,14 @@ export class StreamParser {
 	 * @type {!Promise.<!libjass.ASS>} A promise that will be resolved when the script properties of the ASS script have been parsed from the stream. Styles and events have not necessarily been
 	 * parsed at the point this promise becomes resolved.
 	 */
-	get minimalASS(): promise.Promise<ASS> {
+	get minimalASS(): Promise<ASS> {
 		return this._minimalDeferred.promise;
 	}
 
 	/**
 	 * @type {!Promise.<!libjass.ASS>} A promise that will be resolved when the entire stream has been parsed.
 	 */
-	get ass(): promise.Promise<ASS> {
+	get ass(): Promise<ASS> {
 		return this._deferred.promise;
 	}
 
@@ -301,14 +298,14 @@ export class SrtStreamParser {
 		this._ass.properties.wrappingStyle = 1;
 		this._ass.properties.scaleBorderAndShadow = true;
 
-		var newStyle = new Style(new map.Map([["Name", "Default"]]));
+		var newStyle = new Style(new Map([["Name", "Default"]]));
 		this._ass.styles.set(newStyle.name, newStyle);
 	}
 
 	/**
 	 * @type {!Promise.<!libjass.ASS>} A promise that will be resolved when the entire stream has been parsed.
 	 */
-	get ass(): promise.Promise<ASS> {
+	get ass(): Promise<ASS> {
 		return this._deferred.promise;
 	}
 
@@ -318,7 +315,7 @@ export class SrtStreamParser {
 	private _onNextLine(line: string): void {
 		if (line === null) {
 			if (this._currentDialogueNumber !== null && this._currentDialogueStart !== null && this._currentDialogueEnd !== null && this._currentDialogueText !== null) {
-				this._ass.dialogues.push(new Dialogue(new map.Map([
+				this._ass.dialogues.push(new Dialogue(new Map([
 					["Style", "Default"],
 					["Start", this._currentDialogueStart],
 					["End", this._currentDialogueEnd],
@@ -336,7 +333,7 @@ export class SrtStreamParser {
 
 		if (line === "") {
 			if (this._currentDialogueNumber !== null && this._currentDialogueStart !== null && this._currentDialogueEnd !== null && this._currentDialogueText !== null) {
-				this._ass.dialogues.push(new Dialogue(new map.Map([
+				this._ass.dialogues.push(new Dialogue(new Map([
 					["Style", "Default"],
 					["Start", this._currentDialogueStart],
 					["End", this._currentDialogueEnd],
@@ -422,7 +419,7 @@ export function parseLineIntoTypedTemplate(line: string, formatSpecifier: string
 		value[formatSpecifier.length - 1] = value.slice(formatSpecifier.length - 1).join(",");
 	}
 
-	var template = new map.Map<string, string>();
+	var template = new Map<string, string>();
 	formatSpecifier.forEach((formatKey, index) => {
 		template.set(formatKey, value[index]);
 	});
@@ -441,7 +438,7 @@ export function parse(input: string, rule: string): any {
 	var run = new ParserRun(input, rule);
 
 	if (run.result === null || run.result.end !== input.length) {
-		if (globalSettings.debugMode) {
+		if (debugMode) {
 			console.error("Parse failed. %s %s %o", rule, input, run.result);
 		}
 
@@ -2494,7 +2491,7 @@ makeTagParserFunction("3c", parts.OutlineColor, ParserRun.prototype.parse_color,
 makeTagParserFunction("4a", parts.ShadowAlpha, ParserRun.prototype.parse_alpha, false);
 makeTagParserFunction("4c", parts.ShadowColor, ParserRun.prototype.parse_color, false);
 
-var rules = new map.Map<string, (parent: ParseNode) => ParseNode>();
+var rules = new Map<string, (parent: ParseNode) => ParseNode>();
 Object.keys(ParserRun.prototype).forEach(key => {
 	if (key.indexOf("parse_") === 0 && typeof (<any>ParserRun.prototype)[key] === "function") {
 		rules.set(key.substr("parse_".length), (<any>ParserRun.prototype)[key]);
@@ -2609,7 +2606,7 @@ class ParseNode {
 	}
 }
 
-import webworker = require("./web-worker");
+import * as webworker from "./web-worker";
 
 webworker._registerWorkerCommand(webworker.WorkerCommands.Parse, (parameters: any, response: webworker.WorkerResultCallback): void => {
 	var result: any;
